requirements.txt
Bio
fuzzysearch
ipykernel
jupyter
notebook
Levenshtein
matplotlib
pandas
pybis
PyYAML
seaborn
torch
difflib
#-e .